requirements.txt
# PyTorch with CUDA 12.1 support - compatible versions
torch>=2.0.0
torchvision>=0.15.0
--extra-index-url https://download.pytorch.org/whl/cu121

# Core computer vision
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
opencv-python-headless>=4.8.0
scikit-image>=0.21.0
albumentations>=1.3.0

# Web framework
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# Machine learning & AI
ultralytics>=8.0.0
pandas>=2.0.0
PyYAML>=6.0
tqdm>=4.65.0

# CLIP for zero-shot classification
ftfy>=6.1.1
regex>=2022.1.18
clip-anytorch>=2.5.2

# Additional utilities
psutil>=5.9.0
pathlib2>=2.3.0; python_version < '3.4'
scikit-learn>=1.3.0
Flask-SQLAlchemy>=3.1.1
Flask-JWT-Extended>=4.5.3
email-validator>=2.1.0